+++ backend/src/main/java/hyfive/gachita/path/PathRepository.java
@@ -1,5 +1,6 @@
 package hyfive.gachita.path;
 
+import hyfive.gachita.path.domain.Path;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PathRepository extends JpaRepository<Path, Long> {
