+++ backend/src/main/java/hyfive/gachita/path/domain/PathSegment.java
@@ -1,4 +1,4 @@
-package hyfive.gachita.path;
+package hyfive.gachita.path.domain;
 
 import hyfive.gachita.node.Node;
 import jakarta.persistence.*;
